modified CheatyVision to pull arrays of pre-decoded inputs
Since the python decode lib already exists, probably easier to decode the string on the DS.
Then the DS should send all buttons in an array sorted by stick then button, same for axes.  CheatyVision pulls these from SmartDashboard.

Also, note to Ben: you had been looking earlier in the season for a way to pull Arrays from the SmartDashboard in Java/C++.  Looks like I've found it, but it's not really packaged properly with the SmartDashboard object, but rather with the NetworkTables2 library.  Check out CheatyVision.java for more.