Final (?) Encode fixes
Fixed all remaining Encode.java (and encode.py) bugs

Restructured encoding to put floats in order from the end of the bit string, so that appending zeros from char conversion doesn't screw that up.

Ordered booleans backwards in bitstring so that appending leading zeros from char conversion adds false bools to the END of the bool array, preserving the beginning of the bool array.

Be sure to get the latest encode.py as well.  This file should probably find a place to live in this repo.